Fish shell setup support
`virtualenv` ships with support for `fish` shell and provides `activate.fish` file to activate environment from `fish` shell. Added a comment to use activation file specific to `fish`. Finding it is just a matter of peeking into `bin` directory, and an astute `fish` user may as well find a workaround, but a quick comment saves time.